ref(commands/init): add optional discovery path, move file to commands/
